apps/python-engine/requirements.txt
fastapi
uvicorn
pymupdf
pytesseract
Pillow
python-multipart
yake
underthesea
